Added a keepAlive flag

+++ infrastructure/gce-train/startup.sh
@@ -2,17 +2,24 @@
 
 TRAINER_REPO=$(curl http://metadata.google.internal/computeMetadata/v1/instance/attributes/trainerRepo -H "Metadata-Flavor: Google")
 TRAINER_MODULE=$(curl http://metadata.google.internal/computeMetadata/v1/instance/attributes/trainerModule -H "Metadata-Flavor: Google")
+
 JOB_DIR=$(curl http://metadata.google.internal/computeMetadata/v1/instance/attributes/jobDir -H "Metadata-Flavor: Google")
+
 TRAIN_STEPS=$(curl http://metadata.google.internal/computeMetadata/v1/instance/attributes/trainSteps -H "Metadata-Flavor: Google")
 CHECKPOINT_STEPS=$(curl http://metadata.google.internal/computeMetadata/v1/instance/attributes/checkpointSteps -H "Metadata-Flavor: Google")
 HPARAM1=$(curl http://metadata.google.internal/computeMetadata/v1/instance/attributes/hparam1 -H "Metadata-Flavor: Google")
 HPARAM2=$(curl http://metadata.google.internal/computeMetadata/v1/instance/attributes/hparam2 -H "Metadata-Flavor: Google")
+
+KEEP_ALIVE=$(curl http://metadata.google.internal/computeMetadata/v1/instance/attributes/keepAlive -H "Metadata-Flavor: Google")
+
 
 gcloud source repos clone "$TRAINER_REPO"
 
 cd "$TRAINER_REPO"
 
 git pull origin master
+
+mkdir -p $JOB_DIR
 
 python -m "$TRAINER_MODULE" \
   --job-dir="$JOB_DIR" \
@@ -21,4 +28,6 @@ python -m "$TRAINER_MODULE" \
   --hyperparameter-1="$HPARAM1" \
   --hyperparameter-2="$HPARAM2"
 
-sudo shutdown -h now
+if ! [ $KEEP_ALIVE = "true" ] ; then
+  sudo shutdown -h now
+fi
